improved online status check (#10)
Replaced Check-Online-Status function with one line of code:

Test-Connection -ComputerName $name -Count 1 -Quiet

This returns a Boolean and exits if the machine is not online.

This gives less context if the machine is not online (host could not be found, reply IP address not matching, etc.). The user may use a ping command to see more information.

Empty objects are returned if the -h flag is not selected.