Windows 11 Eligibility Completed (#12)
* (unfinished) Windows 11 Compliance

* Microsoft's Eligibility Script

This is the official script that Microsoft released, to check if hardware of a machine is fit for Windows 11.

* (in progress)

* in progress

* processor check completed

* Added support for -verbose flag
Added support for returning an object with -object flag.

* Added checks for AMD processors.

Removed -verbose handling. Running the script with the object flag and not catching the retun object will effectively do the same thing.

* Added additional comments for clarification.

* removed reference script

* Delete .vscode directory

* Delete bitlocker-progress-bar.ps1